+++ fotoobo/tools/fc/__init__.py
@@ -1,0 +1,7 @@
+"""
+fc tools
+"""
+
+from . import get
+
+__all__ = ["get"]