+++ WHATSNEW.md
@@ -1,6 +1,13 @@
 ### Added
 
+- Add FortiCloud module for cloud asset management
+
 ### Changed
 
-### Removed
+- ems get workgroups now displays group id in table
 
+### Fixed
+
+- Fixed some typos
+
+### Removed